+++ __tests__/cookies.ts
@@ -2,7 +2,7 @@ import puppeteer, { Page } from "puppeteer";
 
 import { defaultPuppeteerBrowserOptions } from "../src/pptr-utils/default";
 import { reportCookieEvents } from "../src/parser";
-import { setupBlacklightInspector } from "../src/inspector";
+import { setupBlacklightInspector } from "../src/inspectors/inspector";
 import { loadEventData } from "../src/utils";
 import { join } from "path";
 import {
