+++ __tests__/js-instrument.ts
@@ -1,7 +1,7 @@
 import puppeteer, { Browser } from "puppeteer";
 import { Global } from "../src/types";
 import { defaultPuppeteerBrowserOptions } from "../src/pptr-utils/default";
-import { setupBlacklightInspector } from "../src/inspector";
+import { setupBlacklightInspector } from "../src/inspectors/inspector";
 declare var global: Global;
 const GETS_AND_SETS = [
   ["prop1", "get", "prop1"],
