+++ src/inspectors/session-recording.ts
@@ -1,5 +1,5 @@
 import { Page } from 'puppeteer';
-import { BlacklightEvent, SESSION_RECORDERS_LIST } from './types';
+import { BlacklightEvent, SESSION_RECORDERS_LIST } from '../types';
 
 export const setupSessionRecordingInspector = async (page: Page, eventDataHandler: (event: BlacklightEvent) => void) => {
     page.on('request', async request => {
